+++ backend/src/main/java/backend/repository/RoomTypeRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.List;
 
-public interface RoomTypeRepository extends JpaRepository<RoomType, Long> {
+public interface RoomTypeRepository extends JpaRepository<RoomType, Integer> {
     List<RoomType> findAllByOrderByTypeNameAsc();
 }
